pc_agent/requirements.txt
# PC Agent 客户端依赖

# 核心依赖
aiohttp>=3.9.0
websockets>=12.0
pyautogui>=0.9.54
pyyaml>=6.0.1
pillow>=10.0.0
pyperclip>=1.8.2

# Windows 专用依赖
pywinauto>=0.6.8; platform_system=="Windows"
pywin32>=306; platform_system=="Windows"

# Mac 专用依赖
pyobjc-framework-Cocoa>=10.0; platform_system=="Darwin"
pyobjc-framework-Quartz>=10.0; platform_system=="Darwin"
pyobjc-framework-ApplicationServices>=10.0; platform_system=="Darwin"
